requirements.txt
# =====================================================================
# Core Dependencies
# =====================================================================
python-dotenv==1.0.0

# =====================================================================
# OpenAI & Embeddings
# =====================================================================
openai==1.54.0
tiktoken==0.8.0

# =====================================================================
# Vector Database
# =====================================================================
qdrant-client==1.12.0

# =====================================================================
# Data Processing
# =====================================================================
numpy==1.26.4
pandas==2.2.0

# =====================================================================
# LangChain & LangGraph (main_langgraph.py용)
# =====================================================================
langchain-core==0.3.0
langchain-openai==0.2.0
langchain-text-splitters==0.3.0

langgraph==0.2.45
langgraph-checkpoint==2.0.2
langgraph-checkpoint-sqlite==2.0.1

# =====================================================================
# OpenAI Agents SDK (main.py용)
# =====================================================================
openai-agents==0.1.4

# =====================================================================
# HTTP Client (Reranker API 호출용)
# =====================================================================
httpx==0.27.0

# =====================================================================
# Data Models
# =====================================================================
pydantic==2.9.0

# =====================================================================
# AWS SDK (S3 통합용 - 선택적)
# =====================================================================
boto3==1.35.0
botocore==1.35.0
